Safari-609.1.11.1/LayoutTests/css3/shapes/shape-outside/values/shape-outside-circle-007-expected.txt


FAIL circle(foo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(10) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20% 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20px, 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20%, 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20px closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20px farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(closest-side 20px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(farthest-side 20px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20% closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(20% farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(closest-side 20%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(farthest-side 20%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(closest-side farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(farthest-side closest-side) - invalid assert_equals: expected (object) null but got (string) ""
